requirements.txt
setuptools>=65.5.1
tensorflow==2.19.0
numpy>=1.26.0
Pillow>=7.1.0,<11.0.0 